refactoring and fix linking visualisation
- renamed "EntryGraph" to "Collection", both because having the entry
prefix made the name stutter a bit "entries.EntryGraph" and because the
data structure itself wasn't a graph, but it could generate a graph.
It's more of a collection of entries.
- renamed "EntryList" to "List" for the same reason as above.
- Links are now not tracked automatically, there's a new FindLinksTo
method on Collection which can find inbound links for an entry. There's
also a new ResolveLink function which uses a Collection to convert a
Link type into an *Entry.
- Fixed collection.Graph method, it's now a directed graph and also
displays titles in the output
- Added more test data to test/