Remove armour and Gzip compression in encryption
Previously, encryption would work like so:

- `entries/` folder is tar-ed and gziped.
- Then compress using gzip again, create ASCII armour, then output

The extra stage of gzip wasn't required and neither was the ASCII armour
because it's stored as a file and not being sent anywhere so it can stay
in a binary format.

This was down to using a package called
https://github.com/jchavannes/go-pgp/ which is a thin wrapper around
`x/crypto/openpgp` which was built for a different use case. This commit
replaces that with a forked version of the package named
albatross-org/go-pgp which removes the additional step of gzipping and
tarring.